Move S3 related code to aws app.
Notable implementation detail: lib.storage_backends.get_storage_manager() now reads from a setting, so that lib.storage_backends doesn't have to import the S3 storage manager.
