+++ project/aws/apps.py
@@ -3,3 +3,7 @@ from django.apps import AppConfig
 
 class AwsConfig(AppConfig):
     name = 'aws'
+
+    def ready(self):
+        # Implicitly connect signal handlers decorated with @receiver.
+        from .tests import signals
